renamed the svgs
